requirements.txt
flask
torch
torchvision
segmentation-models-pytorch
tifffile
pillow